Move cross-switch forwarding wiring to new network/ package
The wireForwarding function integrates simulator/ (NetworkTopology) and
p4runtime/ (P4RuntimeServer/PacketBroker) to implement cross-switch
packet forwarding. It's neither a CLI concern nor a P4Runtime concern —
it's network-level glue that sits above both layers.

New package: fourward.network. Future home for network-level
integrations (e.g., unified network trace observers, multi-switch
trace tree construction for P4Runtime).

Also takes a Map<String, P4RuntimeServer> instead of parallel lists
for a cleaner API.
