+++ cli/BUILD.bazel
@@ -14,6 +14,7 @@ kt_jvm_library(
     visibility = ["//visibility:public"],
     deps = [
         "//e2e_tests/stf:stf_runner",
+        "//network:network_lib",
         "//p4runtime:p4runtime_lib",
         "//simulator:simulator_java_proto",
         "//simulator:simulator_lib",
